refactor(-): :hammer: refatora arquivo rotas para index
