+++ src/helpers/utilities.js
@@ -1241,3 +1241,15 @@ export const ordenarPorLogMaisRecente = (itemA, itemB) => {
 
   return comparaObjetosMoment(dataB, dataA);
 };
+
+export const getDDMMYYYfromDate = (date) => {
+  return `${date.getDate().toString().padStart(2, "0")}/${(date.getMonth() + 1)
+    .toString()
+    .padStart(2, "0")}/${date.getFullYear()}`;
+};
+
+export const getYYYYMMDDfromDate = (date) => {
+  return `${date.getFullYear()}-${(date.getMonth() + 1)
+    .toString()
+    .padStart(2, "0")}-${date.getDate().toString().padStart(2, "0")}`;
+};
